Major refactor in prep for clockwork upgrade.
Should be more maintainable long-term...
Drops the "definition" concept completely. Not sure what I was thinking when I wrote all of that...
Adds php-di/invoker dependency for some magic around actions/filters.
Adds a lot of tests in preparation for the upgrade to clockwork v4.
Bonus: should make the eventual upgrade to v5 (currently in beta) a breeze.
Should also make it simpler to add new features and reorganize logic.